fixed weird error
for some reason message.remove_reaction wasnt a thing and caused the program to not build. But for some reason my home computer allowed it to build even though both computers had the same code so I never noticed this. Adding serenity to cargo.toml seemed to fix it somehow.